make more translations
